compile c in container and base on debian slim
Using the multi-stage build you can compile the source c to a static
binary for use in the actual container.
Since everyone using this repo will be using containers it makes sense
to utilise containers to compile the source c.
Binaries within version control aren't recommended.

This multi-stage build will grab dependencies required to compile the c
and then none of those dependencies that are un-needed afterwards are
copied over.